// src/app/models/detail-measure-control.model.ts
export class DetailMeasureControlModel {
    idDetailMeasureControl: number;
    idHeaderMeasureControl: number;
    dateRegistry: Date;
    currentWeight: number;
    supraPectoral: number;
    subPectoral: number;
    highAbdomen: number;
    waist: number;
    lowerAbdomen: number;
    hip: number;
    thingMeasure: number;
    arm: number;
    triceps: number;
    subEscapular: number;
    suprailiac: number;
    abdominal: number;
    supraspinatus: number;
    thighPercentage: number;
    calf: number;
    grease: number;
    imc: number;
    sistolica: number;
    diastolica: number;
    heartRate: number;
    desiredWeight: number;
    size: number;
    idealWeight: number;
    minWeight: number;
    maxWeight: number;
    observation: string;
    isBlock: boolean;

    constructor(){
        this.idDetailMeasureControl = 0;
        this.idHeaderMeasureControl = 0;
        this.dateRegistry = new Date();
        this.currentWeight = 0;
        this.supraPectoral = 0;
        this.subPectoral = 0;
        this.highAbdomen = 0;
        this.waist = 0;
        this.lowerAbdomen = 0;
        this.hip = 0;
        this.thingMeasure = 0;
        this.arm = 0;
        this.triceps = 0;
        this.subEscapular = 0;
        this.suprailiac = 0;
        this.abdominal = 0;
        this.supraspinatus = 0;
        this.thighPercentage = 0;
        this.calf = 0;
        this.grease = 0;
        this.imc = 0;
        this.sistolica = 0;
        this.diastolica = 0;
        this.heartRate = 0;
        this.desiredWeight = 0;
        this.size = 0;
        this.idealWeight = 0;
        this.minWeight = 0;
        this.maxWeight = 0;
        this.observation = '';
        this.isBlock = true;
    }
}